+++ src/main/java/me/lokka30/phantomworlds/PhantomWorlds.java
@@ -6,18 +6,32 @@ import me.lokka30.microlib.YamlConfigFile;
 import me.lokka30.phantomworlds.commands.phantomworlds.PhantomWorldsCommand;
 import me.lokka30.phantomworlds.managers.FileManager;
 import me.lokka30.phantomworlds.managers.WorldManager;
+import me.lokka30.phantomworlds.misc.CompatibilityChecker;
 import me.lokka30.phantomworlds.misc.UpdateCheckerResult;
 import me.lokka30.phantomworlds.misc.Utils;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import java.io.File;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class PhantomWorlds extends JavaPlugin {
+
+    /**
+     * Contributing code to the plugin?
+     * Feel free to add your name to the list!
+     */
+    @SuppressWarnings("ArraysAsListWithZeroOrOneArgument")
+    public final HashSet<String> contributors = new HashSet<>(Arrays.asList("none"));
+
+    public final String supportedServerVersions = "1.7.x to 1.16.x";
 
     public final FileManager fileManager = new FileManager(this);
     public final WorldManager worldManager = new WorldManager(this);
 
+    public final CompatibilityChecker compatibilityChecker = new CompatibilityChecker();
     public UpdateCheckerResult updateCheckerResult = null;
 
     public YamlConfigFile settings = new YamlConfigFile(this, new File(getDataFolder(), "settings.yml"));
@@ -29,6 +43,7 @@ public class PhantomWorlds extends JavaPlugin {
     public void onEnable() {
         final QuickTimer timer = new QuickTimer();
 
+        checkCompatibility();
         loadFiles();
         loadWorlds();
         registerCommands();
@@ -45,6 +60,21 @@ public class PhantomWorlds extends JavaPlugin {
         /* ... any on-disable content should be put here. nothing for now */
 
         Utils.LOGGER.info("&f~ Shut-down complete. &7Took &b" + timer.getTimer() + "ms");
+    }
+
+    void checkCompatibility() {
+        Utils.LOGGER.info("Checking compatibility with server...");
+
+        compatibilityChecker.checkAll();
+
+        if (compatibilityChecker.incompatibilities.isEmpty()) return;
+
+        AtomicInteger index = new AtomicInteger(1);
+        compatibilityChecker.incompatibilities.forEach(incompatibility -> {
+            Utils.LOGGER.warning("Incompatibility #" + index.getAndIncrement() + " &8(&7Type: " + incompatibility.type + "&8)&7:");
+            Utils.LOGGER.info("&8 &m->&f Reason: &7" + incompatibility.reason);
+            Utils.LOGGER.info("&8 &m->&f Recommendation: &7" + incompatibility.recommendation);
+        });
     }
 
     void loadFiles() {