+++ src/main/java/me/lokka30/phantomworlds/commands/phantomworlds/subcommands/DebugSubcommand.java
@@ -9,26 +9,15 @@ import org.jetbrains.annotations.NotNull;
 import java.util.List;
 
 //TODO
-public class InfoSubcommand implements ISubcommand {
+public class DebugSubcommand implements ISubcommand {
 
     @Override
     public void parseCommand(@NotNull PhantomWorlds main, CommandSender sender, Command cmd, String label, String[] args) {
-        /*
-        TODO
-        Inform the commandsender of the version, authors, contributors,
-        and supported server versions of this version of PhantomWorlds.
-         */
-        sender.sendMessage("Running PhantomWorlds v" + main.getDescription().getVersion());
-        sender.sendMessage("Authors: " + String.join(",", main.getDescription().getAuthors()));
-        sender.sendMessage("Contributors: " + String.join(",", main.contributors));
-        sender.sendMessage("Supported server versions: " + main.supportedServerVersions);
+
     }
 
     @Override
     public List<String> parseTabCompletion(PhantomWorlds main, CommandSender sender, Command cmd, String label, String[] args) {
-        /*
-        TODO
-         */
         return null;
     }
 }