+++ src/main/java/me/lokka30/phantomworlds/managers/WorldManager.java
@@ -6,8 +6,6 @@ import org.bukkit.*;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.IOException;
-
 public class WorldManager {
 
     private final PhantomWorlds main;
@@ -123,27 +121,6 @@ public class WorldManager {
             world.setKeepSpawnInMemory(keepSpawnInMemory);
             world.setPVP(allowPvP);
             world.setDifficulty(difficulty);
-        }
-
-        public void saveToData(PhantomWorlds main) {
-            main.data.getConfig().set("worlds-to-load." + name + ".environment", environment.toString());
-            main.data.getConfig().set("worlds-to-load." + name + ".generateStructures", generateStructures);
-            main.data.getConfig().set("worlds-to-load." + name + ".generator", generator);
-            main.data.getConfig().set("worlds-to-load." + name + ".generatorSettings", generatorSettings);
-            main.data.getConfig().set("worlds-to-load." + name + ".hardcore", hardcore);
-            main.data.getConfig().set("worlds-to-load." + name + ".seed", seed);
-            main.data.getConfig().set("worlds-to-load." + name + ".worldType", worldType);
-            main.data.getConfig().set("worlds-to-load." + name + ".spawnMobs", spawnMobs);
-            main.data.getConfig().set("worlds-to-load." + name + ".spawnAnimals", spawnAnimals);
-            main.data.getConfig().set("worlds-to-load." + name + ".keepSpawnInMemory", keepSpawnInMemory);
-            main.data.getConfig().set("worlds-to-load." + name + ".allowPvP", allowPvP);
-            main.data.getConfig().set("worlds-to-load." + name + ".difficulty", difficulty.toString());
-
-            try {
-                main.data.save();
-            } catch (IOException ex) {
-                ex.printStackTrace();
-            }
         }
     }
 }