+++ src/main/java/me/lokka30/phantomworlds/misc/Utils.java
@@ -52,7 +52,7 @@ public class Utils {
     public static List<String> getPlayersCanSeeList(@NotNull CommandSender sender) {
         List<String> suggestions = new ArrayList<>();
 
-        if (!sender.hasPermission("phantomworlds.vanishbypass") && sender instanceof Player) {
+        if (!sender.hasPermission("phantomworlds.command.see_vanished_users") && sender instanceof Player) {
             Player player = (Player) sender;
             for (Player listedPlayer : Bukkit.getOnlinePlayers()) {
                 if (player.canSee(listedPlayer)) {